+++ src/app/redux/store.ts
@@ -2,7 +2,7 @@ import { configureStore } from '@reduxjs/toolkit';
 import { apiService } from '../../modules/auth/services/apiService';
 import authReducer from '../../modules/auth/slice/authSlice';
 import uiReducer from '../../modules/shared/slice/uiSlice';
-import modalFilterReducer from '../../modules/modals/slice/modalFilterSlice';
+import modalFilterReducer from '../../modules/expenses/slice/modalFilterSlice';
 
 export const store = configureStore({
     reducer: {
