+++ src/modules/auth/components/RegistrationForm.styled.ts
