+++ src/modules/expenses/components/Expenses.tsx
@@ -1,10 +1,10 @@
 import { SystemFilterOutline, ServiceInfoSolid } from '@admiral-ds/icons';
-import { Container, ContainerInfo, InfoMessage, IconInfoSolid } from '../styled/Expenses.styled';
+import { Container, ContainerInfo, InfoMessage, IconInfoSolid } from './Expenses.styled';
 import { ActionBar } from '../../shared/components/ActionBar';
 import { Button, T } from '@admiral-ds/react-ui';
-import { MainContainer } from '../styled/Expenses.styled';
-import { ModalFilter } from '../../modals/components/ModalFilter';
-import { useModalFilter } from '../../modals/hooks/useModalFilter';
+import { MainContainer } from './Expenses.styled';
+import { ModalFilter } from './ModalFilter';
+import { useModalFilter } from '../hooks/useModalFilter';
 
 export const Expenses = () => {
     const { handleOpenModal, isOpen } = useModalFilter();
