+++ src/modules/shared/components/AvatarUser.styled.ts
