+++ src/modules/shared/components/Logo.styled.tsx
