+++ src/modules/shared/layouts/Layout.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from 'react-router-dom';
 import { Header } from '../components/Header';
-import { LayoutStyled } from '../styled/Layout.styled';
+import { LayoutStyled } from './Layout.styled';
 import { BodyGlobal } from '../styled';
 
 export function Layout() {
